applied DRY principle to Struct tests
