+++ src/test/scala/controller/prolog/structs/PathwayDescriptionStructTest.scala
@@ -2,7 +2,8 @@ package controller.prolog.structs
 
 import alice.tuprolog.Var
 import controller.prolog.SqPrologEngine
-import model.nodes.{Pathway, StoryNode}
+import controller.prolog.structs.StructUtil.storyNodeWithAMiddleNodeAndTwoFinalNodes
+import model.nodes.StoryNode
 import org.scalatest.DoNotDiscover
 import specs.FlatTestSpec
 import controller.prolog.util.PrologImplicits._
@@ -12,16 +13,16 @@ import controller.prolog.util.PrologImplicits._
  */
 @DoNotDiscover
 class PathwayDescriptionStructTest extends FlatTestSpec {
-  val finalNode: StoryNode = StoryNode(1, "narrative", None, Set(), List())
-  val pathway: Pathway = Pathway("description", finalNode, None)
-  val startingNode: StoryNode = StoryNode(0, "narrative", None, Set(pathway), List())
-
+  /**
+   * Please @see [[controller.prolog.structs.StructUtil]] for a better understanding of the structure.
+   */
+  val startingNode: StoryNode = storyNodeWithAMiddleNodeAndTwoFinalNodes()
   val engine: SqPrologEngine = SqPrologEngine(startingNode)
 
   "The Prolog engine" should "find the description of the pathway" in {
     val solutions = engine.resolve(PathwayDescriptionStruct(0, 1, new Var()))
     solutions.size shouldEqual 1
-    solutions.head.description shouldEqual "description"
+    solutions.head.description shouldEqual "description(0-1)"
   }
 
   it should "find no solution using wrong ids" in {
