+++ src/test/scala/controller/prolog/structs/StoryNodeStructTest.scala
@@ -2,9 +2,10 @@ package controller.prolog.structs
 
 import alice.tuprolog.Var
 import controller.prolog
+import controller.prolog.structs.StructUtil.storyNodeWithAMiddleNodeAndTwoFinalNodes
 import controller.prolog.{SqPrologEngine, structs}
 import controller.prolog.util.PrologImplicits._
-import model.nodes.{Pathway, StoryNode}
+import model.nodes.StoryNode
 import org.scalatest.DoNotDiscover
 import specs.FlatTestSpec
 
@@ -13,34 +14,31 @@ import specs.FlatTestSpec
  */
 @DoNotDiscover
 class StoryNodeStructTest extends FlatTestSpec {
-
-  val destinationNode: StoryNode = StoryNode(2, "narrative", None, Set.empty, List())
-  val destinationPathway: Pathway = Pathway("description", destinationNode, None)
-  val middleNode: StoryNode = StoryNode(1, "narrative", None, Set(destinationPathway), List())
-  val middlePathway: Pathway = Pathway("description", middleNode, None)
-  val startingNode: StoryNode = StoryNode(0, "narrative", None, Set(middlePathway), List())
-
+  /**
+   * Please @see [[controller.prolog.structs.StructUtil]] for a better understanding of the structure.
+   */
+  val startingNode: StoryNode = storyNodeWithAMiddleNodeAndTwoFinalNodes()
   var engine: SqPrologEngine =  prolog.SqPrologEngine(startingNode)
 
   "The Prolog engine" should "not find solution for a node with index 3" in {
-    engine.resolve(StoryNodeStruct(3, new Var(), new Var())).size shouldEqual 0
+    val invalidIndex = 4
+    engine.resolve(StoryNodeStruct(invalidIndex, new Var(), new Var())).size shouldEqual 0
   }
 
-  it should "find one solution for a node with index 0 and 1 and 2" in {
-    val predicates = for(i <- 0 to 2) yield structs.StoryNodeStruct(i, new Var(), new Var)
+  it should "find one solution for a node with index 0 and 1 and 2 and 3" in {
+    val predicates = for(i <- 0 to 3) yield structs.StoryNodeStruct(i, new Var(), new Var)
     predicates.foreach(p => {
-      engine.resolve(p).head.narrative shouldEqual "narrative"
       engine.resolve(p).size shouldEqual 1
     })
     engine.resolve(predicates(0)).head.pathways.head._1 shouldEqual 1
     engine.resolve(predicates(1)).head.pathways.head._1 shouldEqual 2
     engine.resolve(predicates(2)).head.pathways shouldBe empty
+    engine.resolve(predicates(2)).head.pathways shouldBe empty
   }
 
   it should "find all the solution passing only vars" in {
     val solutions = engine.resolve(structs.StoryNodeStruct(new Var(), new Var(), new Var()))
-    solutions.size shouldEqual 3
-    solutions.foreach(s => s.narrative shouldEqual "narrative")
+    solutions.size shouldEqual 4
   }
 
 }
