+++ src/test/scala/controller/prolog/structs/StructUtil.scala
@@ -1,6 +1,6 @@
 package controller.prolog.structs
 
-import model.nodes.StoryNode
+import model.nodes.{Pathway, StoryNode}
 
 /**
  * Utility methods used to test Prolog structs.
@@ -14,6 +14,20 @@ object StructUtil {
       expected -= s
       res
     })
+  }
+
+  def storyNodeWithAMiddleNodeAndTwoFinalNodes(): StoryNode = {
+    val secondDestinationNode: StoryNode = StoryNode(3, "narrative(3)", None, Set.empty, List())
+    val secondDestinationPathway: Pathway = Pathway("description(1-3)", secondDestinationNode, None)
+    val firstDestinationNode: StoryNode = StoryNode(2, "narrative(2)", None, Set.empty, List())
+    val firstDestinationPathway: Pathway = Pathway("description(1-2)", firstDestinationNode, None)
+    val middleNode: StoryNode = StoryNode(
+      1, "narrative(1)", None,
+      Set(firstDestinationPathway, secondDestinationPathway),
+      List()
+    )
+    val middlePathway: Pathway = Pathway("description(0-1)", middleNode, None)
+    StoryNode(0, "narrative(0)", None, Set(middlePathway), List())
   }
 
 }
